--- .changeset/poor-sheep-tease.md
@@ -1,6 +1,0 @@
----
-'@backstage/plugin-devtools-backend': patch
-'@backstage/plugin-devtools': patch
----
-
-Update devtools information almost real time using signals plugin
