+++ plugins/devtools-backend/api-report.md
@@ -11,7 +11,6 @@ import express from 'express';
 import { ExternalDependency } from '@backstage/plugin-devtools-common';
 import { Logger } from 'winston';
 import { PermissionEvaluator } from '@backstage/plugin-permission-common';
-import { SignalService } from '@backstage/plugin-signals-node';
 
 // @public (undocumented)
 export function createRouter(options: RouterOptions): Promise<express.Router>;
@@ -25,8 +24,6 @@ export class DevToolsBackendApi {
   listExternalDependencyDetails(): Promise<ExternalDependency[]>;
   // (undocumented)
   listInfo(): Promise<DevToolsInfo>;
-  // (undocumented)
-  listResourceUtilization(): Promise<string>;
 }
 
 // @public
@@ -43,7 +40,5 @@ export interface RouterOptions {
   logger: Logger;
   // (undocumented)
   permissions: PermissionEvaluator;
-  // (undocumented)
-  signalService?: SignalService;
 }
 ```
