+++ plugins/devtools-backend/package.json
@@ -38,7 +38,6 @@
     "@backstage/plugin-devtools-common": "workspace:^",
     "@backstage/plugin-permission-common": "workspace:^",
     "@backstage/plugin-permission-node": "workspace:^",
-    "@backstage/plugin-signals-node": "workspace:^",
     "@backstage/types": "workspace:^",
     "@manypkg/get-packages": "^1.1.3",
     "@types/express": "*",
