+++ plugins/devtools-backend/src/api/DevToolsBackendApi.ts
@@ -203,21 +203,17 @@ export class DevToolsBackendApi {
     return configInfo;
   }
 
-  public async listResourceUtilization(): Promise<string> {
+  public async listInfo(): Promise<DevToolsInfo> {
+    const operatingSystem = `${os.hostname()}: ${os.type} ${os.release} - ${
+      os.platform
+    }/${os.arch}`;
     const usedMem = Math.floor((os.totalmem() - os.freemem()) / (1024 * 1024));
-    return `Memory: ${usedMem}/${Math.floor(
+    const resources = `Memory: ${usedMem}/${Math.floor(
       os.totalmem() / (1024 * 1024),
     )}MB - Load: ${os
       .loadavg()
       .map(v => v.toFixed(2))
       .join('/')}`;
-  }
-
-  public async listInfo(): Promise<DevToolsInfo> {
-    const operatingSystem = `${os.hostname()}: ${os.type} ${os.release} - ${
-      os.platform
-    }/${os.arch}`;
-
     const nodeJsVersion = process.version;
 
     /* eslint-disable-next-line no-restricted-syntax */
@@ -251,7 +247,7 @@ export class DevToolsBackendApi {
 
     const info: DevToolsInfo = {
       operatingSystem: operatingSystem ?? 'N/A',
-      resourceUtilization: (await this.listResourceUtilization()) ?? 'N/A',
+      resourceUtilization: resources ?? 'N/A',
       nodeJsVersion: nodeJsVersion ?? 'N/A',
       backstageVersion:
         backstageJson && backstageJson.version ? backstageJson.version : 'N/A',
