+++ plugins/signals-react/package.json
@@ -26,7 +26,7 @@
   "dependencies": {
     "@backstage/core-plugin-api": "workspace:^",
     "@backstage/types": "workspace:^",
-    "@material-ui/core": "^4.9.13"
+    "@material-ui/core": "^4.12.4"
   },
   "peerDependencies": {
     "react": "^16.13.1 || ^17.0.0"
@@ -34,8 +34,8 @@
   "devDependencies": {
     "@backstage/cli": "workspace:^",
     "@backstage/test-utils": "workspace:^",
-    "@testing-library/jest-dom": "^5.10.1",
-    "@testing-library/react": "^12.1.3"
+    "@testing-library/jest-dom": "^6.0.0",
+    "@testing-library/react": "^14.0.0"
   },
   "files": [
     "dist"
