+++ plugins/signals-react/src/hooks/index.ts
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './useSignalApi';
+export * from './useSignal';
